build: adds mobile hero styles
