+++ apps/web/src/containers/LandingPageView/components/hero/components/index.ts
@@ -1,0 +1,2 @@
+export * from './hero-content';
+export * from './hero-stats';
