+++ apps/web/src/containers/LandingPageView/components/hero/images/data.json
